Move core Route wrapper from file to a Router folder
