+++ src/core/components/Router/index.ts
@@ -1,0 +1,1 @@
+export * from './Route';
