+++ src/events/components/EventsRouter.tsx
@@ -1,5 +1,5 @@
 import HttpError from 'core/components/errors/HttpError';
-import Route from 'core/components/Route';
+import { Route } from 'core/components/Router';
 import React from 'react';
 import { Switch } from 'react-router-dom';
 import DetailView from './DetailView';
